Create Custom Scoring

This function geerates a list of scoring settings that can be used as custom scoring setting for calculations. THe resulting list will still need to have a pts_bracket element added to handle points allowed for DSTs. See vignette("scoring_settings") about this.

custom_scoring(...)

Arguments

...
this can be indivdual scoring variables or a list of scoring variables for specific positions. The list will need to be named with a separated list of positions. Note that scoring for passing, kicking, and dst should never be specified within a list.

Examples

# Settings for a standard league:
custom_scoring(pass_yds = 0.04, pass_tds = 4,
               rush_yds = 0.1, rush_tds = 6,
               rec_yds = 0.1, rec_tds = 6)

#> $pass
#> $pass$pass_yds
#> [1] 0.04
#> 
#> $pass$pass_tds
#> [1] 4
#> 
#> 
#> $rush
#> $rush$all_pos
#> [1] TRUE
#> 
#> $rush$rush_yds
#> [1] 0.1
#> 
#> $rush$rush_tds
#> [1] 6
#> 
#> 
#> $rec
#> $rec$all_pos
#> [1] TRUE
#> 
#> $rec$rec_yds
#> [1] 0.1
#> 
#> $rec$rec_tds
#> [1] 6
#> 
#> 

# Settings for a PPR league
custom_scoring(pass_yds = 0.04, pass_tds = 4,
               rush_yds = 0.1, rush_tds = 6,
               rec = 1, rec_yds = 0.1, rec_tds = 6)

#> $pass
#> $pass$pass_yds
#> [1] 0.04
#> 
#> $pass$pass_tds
#> [1] 4
#> 
#> 
#> $rush
#> $rush$all_pos
#> [1] TRUE
#> 
#> $rush$rush_yds
#> [1] 0.1
#> 
#> $rush$rush_tds
#> [1] 6
#> 
#> 
#> $rec
#> $rec$all_pos
#> [1] TRUE
#> 
#> $rec$rec
#> [1] 1
#> 
#> $rec$rec_yds
#> [1] 0.1
#> 
#> $rec$rec_tds
#> [1] 6
#> 
#> 

# Settings for a PPR league with TE premium
custom_scoring(pass_yds = 0.04, pass_tds = 4,
               rush_yds = 0.1, rush_tds = 6,
               "RB, WR, TE" = list(rec = 1, rec_yds = 0.1, rec_tds = 6),
               "TE" = list(rec = 0.5))

#> $pass
#> $pass$pass_yds
#> [1] 0.04
#> 
#> $pass$pass_tds
#> [1] 4
#> 
#> 
#> $rush
#> $rush$all_pos
#> [1] TRUE
#> 
#> $rush$rush_yds
#> [1] 0.1
#> 
#> $rush$rush_tds
#> [1] 6
#> 
#> 
#> $rec
#> $rec$all_pos
#> [1] FALSE
#> 
#> $rec$RB
#> $rec$RB$rec
#> [1] 1
#> 
#> $rec$RB$rec_yds
#> [1] 0.1
#> 
#> $rec$RB$rec_tds
#> [1] 6
#> 
#> 
#> $rec$WR
#> $rec$WR$rec
#> [1] 1
#> 
#> $rec$WR$rec_yds
#> [1] 0.1
#> 
#> $rec$WR$rec_tds
#> [1] 6
#> 
#> 
#> $rec$TE
#> $rec$TE$rec
#> [1] 1.5
#> 
#> $rec$TE$rec_yds
#> [1] 0.1
#> 
#> $rec$TE$rec_tds
#> [1] 6
#> 
#> 
#>